added new classes and changed from float to double

+++ source/IMU.hpp
@@ -16,22 +16,22 @@ namespace rpiScope
 	class IMU_Vector
 	{
 		public:
-			float	X;	//	X,Y,Z
-			float	Y;	//	X,Y,Z
-			float	Z;	//	X,Y,Z
-			float FullScale;
+			double X;	//	X,Y,Z
+			double Y;	//	X,Y,Z
+			double Z;	//	X,Y,Z
+			double FullScale;
 			IMU_Vector();
-			IMU_Vector(float valX, float valY, float valZ, float valScale=1.0);
-			IMU_Vector(int16_t valX, int16_t valY, int16_t valZ, float valScale=1.0);
-			IMU_Vector(int32_t valX, int32_t valY, int32_t valZ, float valScale=1.0);
+			IMU_Vector(double valX, double valY, double valZ, double valScale=1.0);
+			IMU_Vector(int16_t valX, int16_t valY, int16_t valZ, double valScale=1.0);
+			IMU_Vector(int32_t valX, int32_t valY, int32_t valZ, double valScale=1.0);
 			IMU_Vector(IMU_Vector* value);
-			IMU_Vector& set(float valX, float valY, float valZ, float valScale=0.0);
-			IMU_Vector& set(int16_t valX, int16_t valY, int16_t valZ, float valScale=0.0);
-			IMU_Vector& set(int32_t valX, int32_t valY, int32_t valZ, float valScale=0.0);
-			float scaledX(void);
-			float scaledY(void);
-			float scaledZ(void);
-			float Length;
+			IMU_Vector& set(double valX, double valY, double valZ, double valScale=0.0);
+			IMU_Vector& set(int16_t valX, int16_t valY, int16_t valZ, double valScale=0.0);
+			IMU_Vector& set(int32_t valX, int32_t valY, int32_t valZ, double valScale=0.0);
+			double scaledX(void);
+			double scaledY(void);
+			double scaledZ(void);
+			double Length;
 			IMU_Vector& Normalize(void);
 			IMU_Vector* ToEuler(void);
 		protected:
@@ -47,11 +47,11 @@ namespace rpiScope
 			void LPF_resize(size_t LPFValues);
 			IMU_Vector* vector(void);
 			int16_t rawX(void);
-			float scaledX(void);
+			double scaledX(void);
 			int16_t rawY(void);
-			float scaledY(void);
+			double scaledY(void);
 			int16_t rawZ(void);
-			float scaledZ(void);
+			double scaledZ(void);
 			void	Push(int16_t X, int16_t Y, int16_t Z);
 		protected:
 			size_t LPF_MaxValues;
@@ -59,7 +59,7 @@ namespace rpiScope
 			pthread_mutex_t pthread_mutex;
 			void InitMutex(void);
 			void DestroyMutex(void);
-			float FullScale;
+			double FullScale;
 		private:
 	};
 
@@ -75,7 +75,7 @@ namespace rpiScope
 			void PushMagnetometer(int16_t X, int16_t Y, int16_t Z);
 			void PushAcceleration(int16_t X, int16_t Y, int16_t Z);
 			void PushGyroscope(int16_t X, int16_t Y, int16_t Z);
-			void SetFullScale(float gyro, float acc, float mag);
+			void SetFullScale(double gyro, double acc, double mag);
 			//	get calculated values
 			IMU_Vector* Orientation(void);
 		protected:
